Versão 183 - Aplicado fluent builder para criar objetos, a partir de um DTO
Commit (Versão 183 - Aplicado fluent builder para criar objetos, a partir de um DTO)

+++ WebMinimalCarterSwaggerAPI/Modules/ProductsModule.cs
@@ -2,8 +2,11 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.OutputCaching;
 using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
 using System.Net;
+using WebMinimalCarterSwaggerAPI.DTO;
 using WebMinimalCarterSwaggerAPI.Entities;
+using WebMinimalCarterSwaggerAPI.Mappings;
 using WebMinimalCarterSwaggerAPI.Repository.Interfaces;
 
 namespace WebMinimalCarterSwaggerAPI.Modules;
@@ -29,17 +32,19 @@ public sealed class ProductsModule : ICarterModule
 
     [ProducesResponseType((int)HttpStatusCode.Created, Description = "Produto criado com sucesso")]
     [ProducesResponseType((int)HttpStatusCode.BadRequest, Description = "Não foi possivel criar o produto com o payload informado")]
-    private IResult CreateProduct([FromBody] Product product, [FromServices] IProductRepository productsRepo, IOutputCacheStore outputCacheStore)
+    private IResult CreateProduct([FromBody, Required] ProductCreateDTO productCreateDTO, [FromServices] IProductRepository productsRepo, IOutputCacheStore outputCacheStore)
     {
-        if (!product.IsValid()) return Results.BadRequest();
+        var product = ProductMapping.TransformToEntity(productCreateDTO);
+        var productValidation = product.ValidateProductCreation();
+        if (productValidation.Count() > 0) return Results.BadRequest(string.Join(",", productValidation));
         productsRepo.CreateProduct(product);
         outputCacheStore.EvictByTagAsync("produtos", CancellationToken.None);
-        return Results.StatusCode(201);
+        return Results.StatusCode((int)HttpStatusCode.Created);
     }
 
     [ProducesResponseType((int)HttpStatusCode.OK, Description = "Produto obtido com sucesso")]
     [ProducesResponseType((int)HttpStatusCode.BadRequest, Description = "Produto não encontrado para o ID informado")]
-    private IResult GetProductById([FromRoute, Description("ID do produto")] int id, [FromServices] IProductRepository productsRepo)
+    private IResult GetProductById([FromRoute, Description("ID do produto"), Required] int id, [FromServices] IProductRepository productsRepo)
     {
         if (id <= 0) return Results.BadRequest();
         return Results.Ok(productsRepo.GetProductById(id));
@@ -47,9 +52,11 @@ public sealed class ProductsModule : ICarterModule
 
     [ProducesResponseType((int)HttpStatusCode.OK, Description = "Produto atualizado com sucesso")]
     [ProducesResponseType((int)HttpStatusCode.BadRequest, Description = "Não foi possivel atualizar o produto com o payload informado")]
-    private IResult UpdateProduct([FromRoute, Description("ID do produto")] int id, Product product, [FromServices] IProductRepository productRepo, IOutputCacheStore outputCacheStore)
+    private IResult UpdateProduct([FromRoute, Description("ID do produto"), Required] int id, [FromBody, Required] ProductUpdateDTO productUpdateDTO, [FromServices] IProductRepository productRepo, IOutputCacheStore outputCacheStore)
     {
-        if (!product.IsValid()) return Results.BadRequest();
+        var product = ProductMapping.TransformToEntity(productUpdateDTO);
+        var productValidation = product.ValidateProductUpdate();
+        if (productValidation.Count() > 0) return Results.BadRequest(string.Join(",", productValidation));
         productRepo.UpdateProduct(id, product);
         outputCacheStore.EvictByTagAsync("produtos", CancellationToken.None);
         return Results.Ok();
@@ -57,8 +64,9 @@ public sealed class ProductsModule : ICarterModule
 
     [ProducesResponseType((int)HttpStatusCode.OK, Description = "Produto excluido com sucesso", Type = typeof(bool))]
     [ProducesResponseType((int)HttpStatusCode.BadRequest, Description = "Não foi possivel deletar o produto para o ID informado", Type = typeof(bool))]
-    private bool DeleteProduct([FromRoute, Description("ID do produto")] int id, [FromServices] IProductRepository productsRepo, IOutputCacheStore outputCacheStore)
+    private bool DeleteProduct([FromRoute, Description("ID do produto"), Required] int id, [FromServices] IProductRepository productsRepo, IOutputCacheStore outputCacheStore)
     {
+        if (id <= 0) return false;
         var result = productsRepo.DeleteProduct(id);
         outputCacheStore.EvictByTagAsync("produtos", CancellationToken.None);
         return result;
